+++ src/clojure_lsp/feature/file_management.clj
@@ -22,7 +22,7 @@
 (defn ^:private update-findings [db uri new-findings]
   (assoc-in db [:findings (shared/uri->filename uri)] new-findings))
 
-(defn did-open [uri text db]
+(defn did-open [uri text db allow-create-ns]
   (let [settings (get @db :settings {})]
     (when-let [kondo-result (lsp.kondo/run-kondo-on-text! text uri db)]
       (swap! db (fn [state-db]
@@ -32,15 +32,19 @@
                       (update-findings uri (:findings kondo-result))
                       (assoc :kondo-config (:config kondo-result)))))
       (f.diagnostic/async-lint-file! uri db))
-    (when-let [new-ns (and (string/blank? text)
+    (when-let [new-ns (and allow-create-ns
+                           (string/blank? text)
                            (contains? #{:clj :cljs :cljc} (shared/uri->file-type uri))
+                           (not (:processing-work-edit-for-new-files @db))
                            (shared/uri->namespace uri db))]
       (when (get settings :auto-add-ns-to-new-files? true)
         (let [new-text (format "(ns %s)" new-ns)
               changes [{:text-document {:version (get-in @db [:documents uri :v] 0) :uri uri}
                         :edits [{:range (shared/->range {:row 1 :end-row 999999 :col 1 :end-col 999999})
                                  :new-text new-text}]}]]
-          (async/>!! db/edits-chan (f.refactor/client-changes changes db)))))))
+          (async/>!! db/edits-chan (f.refactor/client-changes changes db)))))
+    (when (:processing-work-edit-for-new-files @db)
+      (swap! db assoc :processing-work-edit-for-new-files false))))
 
 (defn ^:private find-changed-var-definitions [old-local-analysis new-local-analysis]
   (let [old-var-defs (filter #(identical? :var-definitions (:bucket %)) old-local-analysis)
@@ -178,5 +182,5 @@
   [uri db]
   (or (get-in @db [:documents uri :text])
       (do
-        (did-open uri (slurp uri) db)
+        (did-open uri (slurp uri) db false)
         (get-in @db [:documents uri :text]))))